Adds KDD datasets.
