+++ ludwig/datasets/configs/kdd_appetency.yaml
@@ -1,0 +1,19 @@
+version: 1.0
+name: kdd_appetency
+download_urls:
+  - https://kdd.org/cupfiles/KDDCupData/2009/orange_small_train.data.zip
+  - https://kdd.org/cupfiles/KDDCupData/2009/orange_small_test.data.zip
+  - https://kdd.org/cupfiles/KDDCupData/2009/orange_small_train_appetency.labels
+  - https://raw.githubusercontent.com/catboost/benchmarks/master/quality_benchmarks/prepare_appetency_churn_upselling/appetency/stratified_train_idx_appetency.txt
+  - https://raw.githubusercontent.com/catboost/benchmarks/master/quality_benchmarks/prepare_appetency_churn_upselling/appetency/stratified_test_idx_appetency.txt
+sha256:
+  orange_small_test.data.zip: 440ac8a350144c14f4d6947c096ad675ee84aa27b4b742071662696e333cec53
+  orange_small_train.data.zip: 31ccb810bdbb71c16e079326443166dc3dfbf73cd358fc4a4ce7440fb1bc6040
+  orange_small_train_appetency.labels: edbfa40e7513804cf25c3f8b3c8f4a6cf5c77116cffc2f87ef770351250a963c
+  stratified_train_idx_appetency.txt: 9c6bf7da6209653e13d9a1d2ef90e4afafe0ecac0eb843c8025816a445c625d9
+  stratified_test_idx_appetency.txt: b80fb8dcf43cd028f4b8affeab65299d580a7e5432ebbe639527dc8177f8764a
+dataset_filenames: orange_small_train.data
+loader: kdd_loader.KDDAppetencyLoader
+description: |
+  The KDD Cup 2009 Appetency dataset.
+  https://www.kdd.org/kdd-cup/view/kdd-cup-2009/Data
