+++ ludwig/datasets/configs/kdd_churn.yaml
@@ -1,0 +1,19 @@
+version: 1.0
+name: kdd_churn
+download_urls:
+  - https://kdd.org/cupfiles/KDDCupData/2009/orange_small_train.data.zip
+  - https://kdd.org/cupfiles/KDDCupData/2009/orange_small_test.data.zip
+  - https://kdd.org/cupfiles/KDDCupData/2009/orange_small_train_churn.labels
+  - https://raw.githubusercontent.com/catboost/benchmarks/master/quality_benchmarks/prepare_appetency_churn_upselling/churn/stratified_train_idx_churn.txt
+  - https://raw.githubusercontent.com/catboost/benchmarks/master/quality_benchmarks/prepare_appetency_churn_upselling/churn/stratified_test_idx_churn.txt
+sha256:
+  orange_small_test.data.zip: 440ac8a350144c14f4d6947c096ad675ee84aa27b4b742071662696e333cec53
+  orange_small_train.data.zip: 31ccb810bdbb71c16e079326443166dc3dfbf73cd358fc4a4ce7440fb1bc6040
+  orange_small_train_churn.labels: fe8891cc574bd55a214514e522a5bed1eec2c3f347a49a36e51620009e7b6f5b
+  stratified_train_idx_churn.txt: 34f9880959ced6f668b25f879fdd388b3826efeca0df03f5a2a5494ce6795406
+  stratified_test_idx_churn.txt: 1675a62cd49c43535eedee3b746f65f8c6a4ebd7f4d0da04e442fd658a408042
+dataset_filenames: orange_small_train.data
+loader: kdd_loader.KDDChurnLoader
+description: |
+  The KDD Cup 2009 Churn dataset.
+  https://www.kdd.org/kdd-cup/view/kdd-cup-2009/Data
