--- ludwig/datasets/kdd_upselling/config.yaml
@@ -1,8 +1,0 @@
-version: 1.0
-download_urls:
-  - "https://kdd.org/cupfiles/KDDCupData/2009/orange_small_train.data.zip"
-  - "https://kdd.org/cupfiles/KDDCupData/2009/orange_small_test.data.zip"
-  - "https://kdd.org/cupfiles/KDDCupData/2009/orange_small_train_upselling.labels"
-  - "https://raw.githubusercontent.com/catboost/benchmarks/master/quality_benchmarks/prepare_appetency_churn_upselling/upselling/stratified_train_idx_upselling.txt"
-  - "https://raw.githubusercontent.com/catboost/benchmarks/master/quality_benchmarks/prepare_appetency_churn_upselling/upselling/stratified_test_idx_upselling.txt"
-csv_filename: kdd_upselling.csv
